requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
dnspython==1.16.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
packaging==20.3
pluggy==0.13.1
protobuf==3.6.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dotenv==0.12.0
python-env==1.0.0
redis==3.4.1
requests==2.23.0
six==1.14.0
toml==0.10.0
tox==3.14.6
urllib3==1.25.8
virtualenv==20.0.14
wcwidth==0.1.8
zipp==3.1.0
